move LoggingDiary to api
