+++ api/src/main/java/com/microsoft/gctoolkit/jvm/SupportedFlags.java
@@ -1,6 +1,6 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
-package com.microsoft.gctoolkit.parser.jvm;
+package com.microsoft.gctoolkit.jvm;
 
 public enum SupportedFlags {
     APPLICATION_STOPPED_TIME,                   //  0
