+++ api/src/main/java/com/microsoft/gctoolkit/parser/datatype/TripleState.java
