+++ parser/src/main/java/com/microsoft/gctoolkit/parser/PreUnifiedG1GCParser.java
@@ -28,7 +28,7 @@ import com.microsoft.gctoolkit.event.jvm.JVMEvent;
 import com.microsoft.gctoolkit.event.jvm.JVMTermination;
 import com.microsoft.gctoolkit.time.DateTimeStamp;
 import com.microsoft.gctoolkit.parser.collection.MRUQueue;
-import com.microsoft.gctoolkit.parser.jvm.LoggingDiary;
+import com.microsoft.gctoolkit.jvm.LoggingDiary;
 
 import java.util.AbstractMap;
 import java.util.concurrent.ConcurrentLinkedQueue;
