+++ parser/src/main/java/com/microsoft/gctoolkit/parser/vmops/SafepointParser.java
@@ -6,7 +6,7 @@ import com.microsoft.gctoolkit.event.jvm.JVMTermination;
 import com.microsoft.gctoolkit.event.jvm.Safepoint;
 import com.microsoft.gctoolkit.parser.JVMEventConsumer;
 import com.microsoft.gctoolkit.parser.PreUnifiedGCLogParser;
-import com.microsoft.gctoolkit.parser.jvm.LoggingDiary;
+import com.microsoft.gctoolkit.jvm.LoggingDiary;
 
 
 public class SafepointParser extends PreUnifiedGCLogParser implements SafepointPatterns {
