+++ parser/src/test/java/com/microsoft/gctoolkit/parser/diary/LogDiaryTest.java
@@ -5,11 +5,11 @@ package com.microsoft.gctoolkit.parser.diary;
 import com.microsoft.gctoolkit.parser.TestLogFile;
 import com.microsoft.gctoolkit.io.GCLogFile;
 import com.microsoft.gctoolkit.io.SingleGCLogFile;
+import com.microsoft.gctoolkit.jvm.SupportedFlags;
 import com.microsoft.gctoolkit.parser.jvm.JVMConfiguration;
-import com.microsoft.gctoolkit.parser.jvm.LoggingDiary;
+import com.microsoft.gctoolkit.jvm.LoggingDiary;
 import com.microsoft.gctoolkit.parser.jvm.PreUnifiedJVMConfiguration;
 import com.microsoft.gctoolkit.parser.jvm.UnifiedJVMConfiguration;
-import com.microsoft.gctoolkit.parser.jvm.*;
 
 import java.io.File;
 import java.io.IOException;
