+++ parser/src/test/java/com/microsoft/gctoolkit/parser/patterns/UnifiedGenerationalEventsTest.java
@@ -19,7 +19,7 @@ import com.microsoft.gctoolkit.event.generational.ParNew;
 import com.microsoft.gctoolkit.event.jvm.JVMEvent;
 import com.microsoft.gctoolkit.parser.UnifiedGenerationalParser;
 import com.microsoft.gctoolkit.parser.jvm.UnifiedJVMConfiguration;
-import com.microsoft.gctoolkit.parser.jvm.LoggingDiary;
+import com.microsoft.gctoolkit.jvm.LoggingDiary;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
