+++ vertx/src/main/java/com/microsoft/gctoolkit/vertx/jvm/GCToolkitVertxParametersForUnifiedLogs.java
@@ -11,7 +11,7 @@ import com.microsoft.gctoolkit.parser.UnifiedJVMEventParser;
 import com.microsoft.gctoolkit.parser.UnifiedSurvivorMemoryPoolParser;
 import com.microsoft.gctoolkit.parser.ZGCParser;
 import com.microsoft.gctoolkit.parser.jvm.JVMConfiguration;
-import com.microsoft.gctoolkit.parser.jvm.LoggingDiary;
+import com.microsoft.gctoolkit.jvm.LoggingDiary;
 import com.microsoft.gctoolkit.vertx.GCToolkitVertx;
 import com.microsoft.gctoolkit.vertx.aggregator.AggregatorVerticle;
